requirements.txt
asgiref==3.5.2
certifi==2022.6.15.1
charset-normalizer==2.1.1
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==4.1.1
djangorestframework==3.13.1
Faker==14.2.0
httpie==3.2.1
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
openapi-codec==1.3.2
pkg_resources==0.0.0
Pygments==2.13.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
requests-toolbelt==0.9.1
rich==12.5.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.12
